JITServer AOT cache deserializer implementation
The JITServerAOTDeserializer class is used at the JIT client to
deserialize cached AOT methods received from the JITServer. It updates
each SCC offset in relocation and validation records stored in method
relocation data such that it points to the correct entity in the
local SCC, which is looked up and validated using the information in
AOT serialization records attached to the serialized AOT method.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>