Compare handles when merging MCS target operands
...in InterpreterEmulator. This test looks like it was intended to check
that both target MethodHandles are the same object, rather than simply
that both have a nonzero known object index (which only tells us that
the objects are non-null).

It shouldn't be strictly necessary for the target MethodHandles to be
the same. In case they differ, it should be fine to continue with the
more recent value. However, determining which value is more recent would
take additional tracking, and the values will only differ when the
MutableCallSite is updated during the inlining pass, i.e. very rarely.