Add support for delayed start of JITServer
When JITServer clients are run in the containerized environment,
they are likely allocated a limited amount of memory that's sufficient
to run the application and support JITServer compilation, but not large
enough to perform expensive compilations locally.

If a server is temporarily unavailable, all compilations will be done
locally and we will see many failures due to low available memory.
If a compilation failed early due to insufficient memory, there will be
no recompilation attempts leading to bad performance even if server
becomes available or memory situation improves.

This commit addresses the above issue in 2 steps:
1. If the client is running low on memory, downgrade compilations to
cold, because they are typically much cheaper, making them less likely
to fail.
2. Once the downgraded compilation finishes,
add an upgrade request into the Low Priority Queue (LPQ).
Requests from the LPQ are processed when the main queue has no compilation requests left.
If by the time we get to processing requests from the LPQ the server is available, client will
send upgrade requests to the server, potentially regaining all
of the performance lost.

Closes: #12375

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>