Fix memory leaks when using fast client session destruction
PR #12458 introduced fast client session destruction: most of the
destructor code can be skipped since most data in the client session
is allocated with the per-client persistent allocator and is freed in
bulk when the per-client allocator is destroyed. Any objects allocated
globally (not with the per-client allocator) still need to be freed.
This commit fixes missing deallocation of std::string objects in the
client session (which are allocated using the global malloc heap).

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>