Account for array header alignment for multianewarray evaluator
The array header size recently increased from 16 bytes to 24 bytes on
64-bit platforms.  With a 16 byte header size, inline code generated by
the JIT to allocate a multidimensional array on the thread-local heap
didn't need to account for alignment of the array headers, as they would
be aligned for any shift compression amount.  With a size of 24 bytes,
headers will not be automatically aligned for a shift compression
amount of four, which requires a heap alignment of 16 bytes.

Adjusted multianewarray evaluators for x and z platforms to account for
array header alignment requirements.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>