Fix potential Segmentation Error due to a NULL hotFieldPoolMutex
Potentially fix segmentation error with vmState=0x000565ff due to a NULL
hotFieldPoolMutex. I have been unable to reproduce the issue, however, from analyzing
the code, this patch addresses the corner case where the hotFieldPool for the
classloader is successfully created while the hotFieldPoolMutex for the classloader
is not successfully created.

Related to: https://github.com/eclipse-openj9/openj9/issues/12782

Signed-off-by: Jonathan Oommen <jon.oommen@gmail.com>