Checkout directly to OPENJSK_SHA if passed
On a Docker based compile, SCM checkout runs
using git from the host. Git checkout uses Git
from the container. This seems to produce an
incompatilibilty between the git metadata. When
'git checkout' runs (in the container) it can't
seem to process correctly and fails "fatal: reference
is not a tree: <sha>". Having the scm step checkout
directly to the SHA if it is passed should eliminate
the need to run the extra checkout. Note that
OPENJDK_SHA should be passed in 100% of the builds.
It would only not be passed if someone was launching
a build job direcly without filling in the param.

Test cases
- OpenJ9 PR
- OpenJDK PR
- OpenJ9 PR depends OpenJDK PR
- Non-PR build with OpenJDK_SHA passed
- Non-PR build without OPENJDK_SHA passed

Fixes #12881

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>