Add additional check for romClass in iterateStackTrace()
Always make sure the ramClass is the correct one in iterateStackTrace().
We probe the bootstrap class loader and then appClass loader to find the
corresponding ramClass of a romClass. However, the correct class could
be loaded by another class loader. In this case, the returned ramClass
is not correct, which should not be used. 

Fixes #12915

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>